# ecommerce/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Vendor AI E-commerce API
  version: 1.0.0
  description: REST API for the multi-vendor AI e-commerce platform.
paths:
...
  /api/v1/admin/products/{product_id}/:
    delete:
      operationId: admin_products_delete
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDeleteResponse'
          description: ''
  /api/v1/admin/products/{product_id}/approve/:
    post:
      operationId: admin_products_approve
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductResponse'
          description: ''
  /api/v1/admin/products/{product_id}/hide/:
    post:
      operationId: admin_products_hide
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductResponse'
          description: ''
  /api/v1/admin/products/{product_id}/reject/:
    post:
      operationId: admin_products_reject
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminProductReject'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminProductReject'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminProductReject'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductResponse'
          description: ''
  /api/v1/admin/products/pending/:
    get:
      operationId: admin_products_pending
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductListResponse'
          description: ''
...
  /api/v1/products/:
    get:
      operationId: public_products_list
      parameters:
      - in: query
        name: brand_id
        schema:
          type: string
          format: uuid
      - in: query
        name: category_id
        schema:
          type: string
          format: uuid
      - in: query
        name: max_price
        schema:
          type: string
          format: decimal
          pattern: ^-?\d{0,18}(?:\.\d{0,0})?$
      - in: query
        name: min_price
        schema:
          type: string
          format: decimal
          pattern: ^-?\d{0,18}(?:\.\d{0,0})?$
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
          maxLength: 255
          minLength: 1
      - in: query
        name: sort
        schema:
          enum:
          - created_at
          - -created_at
          - price
          - -price
          - sold_count
          - -sold_count
          - rating
          - -rating
          type: string
          default: -created_at
          minLength: 1
        description: |-
          * `created_at` - created_at
          * `-created_at` - -created_at
          * `price` - price
          * `-price` - -price
          * `sold_count` - sold_count
          * `-sold_count` - -sold_count
          * `rating` - rating
          * `-rating` - -rating
      tags:
      - v1
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPublicProductListResponseList'
          description: ''
  /api/v1/products/{slug}/:
    get:
      operationId: public_products_retrieve
      parameters:
      - in: query
        name: shop_slug
        schema:
          type: string
        description: Bắt buộc khi nhiều shop có cùng product slug.
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProductResponse'
          description: ''
...
  /api/v1/seller/attributes/:
    get:
      operationId: seller_attributes_list
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeListResponse'
          description: ''
  /api/v1/seller/products/:
    get:
      operationId: seller_products_list
      parameters:
      - in: query
        name: category_id
        schema:
          type: string
          format: uuid
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
          maxLength: 255
          minLength: 1
      - in: query
        name: sort
        schema:
          enum:
          - created_at
          - -created_at
          - price
          - -price
          - sold_count
          - -sold_count
          - rating
          - -rating
          type: string
          default: -created_at
          minLength: 1
        description: |-
          * `created_at` - created_at
          * `-created_at` - -created_at
          * `price` - price
          * `-price` - -price
          * `sold_count` - sold_count
          * `-sold_count` - -sold_count
          * `rating` - rating
          * `-rating` - -rating
      - in: query
        name: status
        schema:
          enum:
          - draft
          - pending_review
          - approved
          - rejected
          - hidden
          - suspended
          type: string
          minLength: 1
        description: |-
          * `draft` - Bản nháp
          * `pending_review` - Chờ duyệt
          * `approved` - Đã duyệt
          * `rejected` - Bị từ chối
          * `hidden` - Đã ẩn
          * `suspended` - Tạm ngưng
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSellerProductListResponseList'
          description: ''
    post:
      operationId: seller_products_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerProductCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SellerProductCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SellerProductCreate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerProductResponse'
          description: ''
  /api/v1/seller/products/{product_id}/:
    get:
      operationId: seller_products_retrieve
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerProductResponse'
          description: ''
    patch:
      operationId: seller_products_update
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedSellerProductUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedSellerProductUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedSellerProductUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerProductResponse'
          description: ''
    delete:
      operationId: seller_products_delete
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDeleteResponse'
          description: ''
  /api/v1/seller/products/{product_id}/media/:
    post:
      operationId: seller_products_media_upload
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaUpload'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MediaUpload'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MediaUpload'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaResponse'
          description: ''
  /api/v1/seller/products/{product_id}/media/{media_id}/:
    delete:
      operationId: seller_products_media_delete
      parameters:
      - in: path
        name: media_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDeleteResponse'
          description: ''
  /api/v1/seller/products/{product_id}/media/reorder/:
    post:
      operationId: seller_products_media_reorder
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaReorder'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MediaReorder'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MediaReorder'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerProductResponse'
          description: ''
  /api/v1/seller/products/{product_id}/submit/:
    post:
      operationId: seller_products_submit
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerProductResponse'
          description: ''
  /api/v1/seller/products/{product_id}/variants/{variant_id}/:
    patch:
      operationId: seller_products_variants_update
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: variant_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedVariantUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedVariantUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedVariantUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantResponse'
          description: ''
  /api/v1/seller/products/{product_id}/variants/generate/:
    post:
      operationId: seller_products_variants_generate
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VariantGenerate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VariantGenerate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VariantGenerate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantListResponse'
          description: ''
...
components:
  schemas:
...
    AdminProductListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminProductList'
        meta:
          $ref: '#/components/schemas/ProductPaginationMeta'
      required:
      - data
      - message
      - meta
      - success
    AdminProductReject:
      type: object
      properties:
        rejection_reason:
          type: string
          maxLength: 2000
      required:
      - rejection_reason
    AdminProductResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/AdminProductList'
      required:
      - data
      - message
      - success
...
    AttributeListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/AttributeList'
        meta:
          $ref: '#/components/schemas/ProductPaginationMeta'
      required:
      - data
      - message
      - meta
      - success
...
    PaginatedPublicProductListResponseList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/PublicProductListResponse'
    PaginatedSellerProductListResponseList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/SellerProductListResponse'
...
    ProductDeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          additionalProperties: {}
      required:
      - data
      - message
      - success
...
    PublicProductResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/PublicProductDetail'
      required:
      - data
      - message
      - success
...
    SellerProductResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/SellerProductDetail'
      required:
      - data
      - message
      - success
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT